requirements.txt
openai-whisper\nffmpeg-python
